+++ src/components/sbb-datepicker-previous-day/sbb-datepicker-previous-day.e2e.ts
@@ -25,7 +25,7 @@ describe('sbb-datepicker-previous-day', () => {
       const input: E2EElement = await page.find('input');
       await page.waitForChanges();
       expect(element).toHaveClass('hydrated');
-      expect(await input.getProperty('value')).toEqual('01-01-2023');
+      expect(await input.getProperty('value')).toEqual('01.01.2023');
 
       const changeSpy = await input.spyOnEvent('change');
       await element.click();
@@ -58,7 +58,7 @@ describe('sbb-datepicker-previous-day', () => {
     });
 
     it('click', async () => {
-      expect(await input.getProperty('value')).toEqual('20-01-2023');
+      expect(await input.getProperty('value')).toEqual('20.01.2023');
       const changeSpy = await input.spyOnEvent('change');
       await element.click();
       await waitForCondition(() => changeSpy.events.length === 1);
@@ -78,7 +78,7 @@ describe('sbb-datepicker-previous-day', () => {
       input = await page.find('input');
       await page.waitForChanges();
 
-      expect(await input.getProperty('value')).toEqual('20-01-2023');
+      expect(await input.getProperty('value')).toEqual('20.01.2023');
       expect(
         await page.evaluate(() =>
           document.querySelector('sbb-datepicker-previous-day').getAttribute('disabled')
@@ -87,18 +87,18 @@ describe('sbb-datepicker-previous-day', () => {
 
       await element.click();
       await page.waitForChanges();
-      expect(await input.getProperty('value')).toEqual('20-01-2023');
+      expect(await input.getProperty('value')).toEqual('20.01.2023');
     });
 
     it('disabled due disabled picker', async () => {
-      expect(await input.getProperty('value')).toEqual('20-01-2023');
+      expect(await input.getProperty('value')).toEqual('20.01.2023');
       await page.evaluate(() => document.querySelector('input').setAttribute('disabled', ''));
       await page.waitForChanges();
 
       expect(element.getAttribute('disabled')).toEqual('');
       await element.click();
       await page.waitForChanges();
-      expect(await input.getProperty('value')).toEqual('20-01-2023');
+      expect(await input.getProperty('value')).toEqual('20.01.2023');
     });
   });
 });