+++ src/components/sbb-selection-panel/sbb-selection-panel.stories.jsx
@@ -297,6 +297,54 @@ const TicketsOptionsExampleTemplate = ({ checkedInput, disabledInput, ...args })
   </sbb-checkbox-group>
 );
 
+const NestedRadioTemplate = ({ disabledInput, ...args }) => (
+  <sbb-radio-button-group orientation="vertical" horizontal-from="large">
+    <sbb-selection-panel {...args}>
+      <sbb-radio-button value="mainoption1" checked={checkedInput}>
+        Main Option 1
+      </sbb-radio-button>
+      <sbb-radio-button-group orientation="vertical" slot="content">
+        <sbb-radio-button value="suboption1">Suboption 1</sbb-radio-button>
+        <sbb-radio-button value="suboption2">Suboption 2</sbb-radio-button>
+      </sbb-radio-button-group>
+    </sbb-selection-panel>
+
+    <sbb-selection-panel {...args}>
+      <sbb-radio-button value="mainoption2" disabled={disabledInput}>
+        Main Option 2
+      </sbb-radio-button>
+      <sbb-radio-button-group orientation="vertical" slot="content">
+        <sbb-radio-button value="suboption1">Suboption 1</sbb-radio-button>
+        <sbb-radio-button value="suboption2">Suboption 2</sbb-radio-button>
+      </sbb-radio-button-group>
+    </sbb-selection-panel>
+  </sbb-radio-button-group>
+);
+
+const NestedCheckboxTemplate = ({ disabledInput, ...args }) => (
+  <sbb-checkbox-group orientation="vertical" horizontal-from="large">
+    <sbb-selection-panel {...args}>
+      <sbb-checkbox value="mainoption1" checked={checkedInput}>
+        Main Option 1
+      </sbb-checkbox>
+      <sbb-checkbox-group orientation="vertical" slot="content">
+        <sbb-checkbox value="suboption1">Suboption 1</sbb-checkbox>
+        <sbb-checkbox value="suboption2">Suboption 2</sbb-checkbox>
+      </sbb-checkbox-group>
+    </sbb-selection-panel>
+
+    <sbb-selection-panel {...args}>
+      <sbb-checkbox value="mainoption2" disabled={disabledInput}>
+        Main Option 2
+      </sbb-checkbox>
+      <sbb-checkbox-group orientation="vertical" slot="content">
+        <sbb-checkbox value="suboption1">Suboption 1</sbb-checkbox>
+        <sbb-checkbox value="suboption2">Suboption 2</sbb-checkbox>
+      </sbb-checkbox-group>
+    </sbb-selection-panel>
+  </sbb-checkbox-group>
+);
+
 const WithCheckboxesErrorMessageTemplate = ({ checkedInput, disabledInput, ...args }) => {
   const sbbFormError = <sbb-form-error slot="error">This is a required field.</sbb-form-error>;
 
@@ -523,6 +571,14 @@ WithNoContent.args = { ...basicArgs, checkedInput: true };
 export const TicketsOptionsExample = TicketsOptionsExampleTemplate.bind({});
 TicketsOptionsExample.argTypes = basicArgTypes;
 TicketsOptionsExample.args = { ...basicArgs, checkedInput: true };
+
+export const NestedRadios = NestedRadioTemplate.bind({});
+NestedRadios.argTypes = basicArgTypes;
+NestedRadios.args = { ...basicArgs };
+
+export const NestedCheckboxes = NestedCheckboxTemplate.bind({});
+NestedCheckboxes.argTypes = basicArgTypes;
+NestedCheckboxes.args = { ...basicArgs };
 
 export default {
   decorators: [